Edits for deployment & fixed dynamic routes
