+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "next dev -p 5000",
     "build": "next build",
-    "start": "next start",
+    "start": "next start -p 5000",
     "lint": "next lint"
   },
   "author": "jiahui,amalie,agatha,talia,syakira",
